update README and do some cleanup
