--- src/app/auth/error/page.tsx
@@ -1,5 +1,0 @@
-'use client';
-
-export default function ErrorPage() {
-  return <p>Sorry, something went wrong</p>
-}
